// src/lib/types/generated/CharacterPackWithSounds.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CharacterEventSound } from "./CharacterEventSound";

export type CharacterPackWithSounds = { sounds: Array<CharacterEventSound>, id: string, name: string, display_name: string, language: string | null, avatar_path: string | null, is_bundled: boolean, is_enabled: boolean, is_complete: boolean, created_at: string, };